# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/account/verification:
    post:
      operationId: sendAccountVerification
      description: Send a one-time verification link to the authenticated Console account. A previous unused link is invalidated.
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthVerificationRequest' } } } }
      responses:
        '204': { description: Verification message accepted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    put:
      operationId: confirmAccountVerification
      description: Consume a one-time Console email verification token.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenRequest' } } } }
      responses:
        '200': { description: Verified account, content: { application/json: { schema: { $ref: '#/components/schemas/AccountResponse' } } } }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/account/recovery:
    post:
      operationId: createAccountRecovery
      description: Request a password reset email. The accepted response is identical whether the address exists.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordRecoveryRequest' } } } }
      responses:
        '202': { description: Recovery request accepted, content: { application/json: { schema: { $ref: '#/components/schemas/AcceptedResponse' } } } }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: confirmAccountRecovery
      description: Replace the Console password with a valid one-time recovery token and revoke all existing sessions.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } } }
      responses:
        '204': { description: Password updated and sessions revoked }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes are users.read/write, databases.read/write, storage.read/write, functions.read/write, sites.read/write, webhooks.read/write, and realtime.read; never send this header through the Console browser bridge. }
...
  responses:
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    RateLimited: { description: Public Auth rate limit exceeded, headers: { Retry-After: { description: Seconds until another attempt may be made, schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ServiceUnavailable: { description: Public Auth protection or a required dependency is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    AuthVerificationRequest: { type: object, additionalProperties: false, description: Empty JSON object by default; an optional trusted url may receive the tokenized link., properties: { url: { type: string, format: uri, maxLength: 2048 } } }
    AuthTokenRequest: { type: object, additionalProperties: false, anyOf: [{ required: [token] }, { required: [secret] }], properties: { token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$', writeOnly: true }, secret: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$', writeOnly: true }, user_id: { type: string, format: uuid, description: Compatibility field; the opaque token is authoritative. } } }
    PasswordRecoveryRequest: { type: object, additionalProperties: false, required: [email], properties: { email: { type: string, format: email, maxLength: 320 }, url: { type: string, format: uri, maxLength: 2048 } } }
    PasswordResetRequest: { type: object, additionalProperties: false, required: [password], anyOf: [{ required: [token] }, { required: [secret] }], properties: { token: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$', writeOnly: true }, secret: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$', writeOnly: true }, user_id: { type: string, format: uuid, description: Compatibility field; the opaque token is authoritative. }, password: { type: string, minLength: 12, maxLength: 256, writeOnly: true } } }
...
    Account: { type: object, required: [id, email, email_verified, created_at], properties: { id: { type: string, format: uuid }, email: { type: string, format: email }, email_verified: { type: boolean }, created_at: { type: string, format: date-time } } }
...
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
    AccountResponse: { type: object, required: [account], properties: { account: { $ref: '#/components/schemas/Account' } } }
...
    AcceptedResponse: { type: object, additionalProperties: false, required: [status], properties: { status: { type: string, enum: [accepted] } } }
...